+++ libitina/frc_2018.py
@@ -4,4 +4,10 @@
 # Licensed under Apache 2.0 License
 #
 # Provides the functions needed for the robot specific to the FRC 2018 competition
-from frc_functions import *
+from libitina.frc_functions import *
+
+
+def run_automatically():
+    print("Running autonomous code...")
+    move_forward(3)
+    stop()
