+++ main.py
@@ -4,11 +4,11 @@
 # Licensed under Apache 2.0 License
 #
 # Calls all functions needed and runs the main program
-from frc_functions import *
+from libitina.frc_2018 import *
 
 
 # Start the main process
 if __name__ == '__main__':
     # Put some main code here!
     print("Hi, Monika.")
-    move_forward(5)
+    run_automatically()
