+++ settings.py
@@ -5,6 +5,7 @@
 #
 # Offers a dialog for changing settings such as position and alliance
 from PyQt5 import *
+from libitina.libitina import Libitina
 
 whichAlliance = 0
 # 0 indicates 'red', 1 indicates blue
@@ -12,3 +13,5 @@ whichAlliance = 0
 if __name__ == '__main__':
     # TODO: Create the GUI for settings.
     print("Hi, Monika.")
+    robot = Libitina()
+    robot.enable_third_eye()
